course/module approval added

+++ dbOperations/HoI_approve_course_db.php
@@ -1,12 +1,5 @@
 <?php
 
-    $link = mysqli_connect("localhost", "root", "", "courses_details");
-
-
-    if($link === false){
-        die("ERROR: Could not connect. " . mysqli_connect_error());
-    }
-
     $course_name =$_POST["course_name"] ;
     $student_count = $_POST["student_count"];
     $duration = $_POST["duration"];
@@ -17,42 +10,63 @@
     $medium = $_POST["medium"];
     $required_qualification = $_POST["required_qualification"];
     $course_id = $_POST["course_id"];
-
     $id = "C"."$course_id";
 
 
-    $sql = "INSERT INTO course_details (course_id,trade,course_name,course_type, type, accredit_level, duration,   medium,required_qualification,student_count) VALUES ('$id', '$trade', '$course_name', '$course_type', '$type', '$accredit_level', '$duration', '$medium', '$required_qualification', '$student_count')";
+
+    $link = mysqli_connect("localhost", "root", "", "courses_details");
+    if($link === false){
+        die("ERROR: Could not connect. " . mysqli_connect_error());
+    }
+    $sql = "INSERT INTO course_details (course_id,trade,course_name,course_type, type, accredit_level, duration,   medium,required_qualification,student_count) VALUES
+    ('$id', '$trade', '$course_name', '$course_type', '$type', '$accredit_level', '$duration', '$medium', '$required_qualification', '$student_count')";
     if(mysqli_query($link, $sql)){
         echo "Records inserted successfully.";
 
 
         $link2 = mysqli_connect("localhost", "root", "", "course_info");
-
-
         if($link2 === false){
             die("ERROR: Could not connect. " . mysqli_connect_error());
         }
 
-        $course_id =  $_POST["course_id"];
+        $sql4 = "SELECT module_name, description, lecturer FROM modules where course_id = '$course_id'";
+
+        $resultset = mysqli_query($link2, $sql4);
+        mysqli_close($link2);
+
+
+
+
+
+        if($resultset){
+          $count = 1;
+          while($row = mysqli_fetch_array($resultset)){
+              $module_name= $row[0];
+              echo $module_name;
+              $description = $row[1];
+              echo $description;
+              $lecturer = $row[2];
+              echo $lecturer;
+              $module_id = "$course_id"."m"."$count";
+              $sql4 = "INSERT INTO module_details (module_id,module_name,no_of_students, lecturer, description) VALUES ('$module_id', '$module_name', '$student_count', '$lecturer', '$description')";
+              if(mysqli_query($link, $sql4)){
+                echo "module approved";
+              } else{
+                   echo "ERROR: Could not able to execute $sql4. " . mysqli_error($link);
+              }
+              $count = $count +1;
+
+          }
+
+        $link2 = mysqli_connect("localhost", "root", "", "course_info");
+        if($link2 === false){
+            die("ERROR: Could not connect. " . mysqli_connect_error());
+        }
         $sql2 = "delete from modules where course_id = $course_id";
-
-        $query = "SELECT id, id_key, name, file_name, file_path, description, added_date, remove_date FROM temporary_news WHERE id = '$id'";
-        $result = mysqli_query($link,$query);
-        $data = mysqli_fetch_array($result);
-        $id_key = $data[1];
-        $name = $data[2];
-        $file_name = $data[3];
-        $file_path = $data[4];
-        $description = $data[5];
-        $added_date = $data[6];
-        $remove_date = $data[7];
-
         if(mysqli_query($link2, $sql2)){
-
             $sql3 = "delete from courses where id = $course_id";
-
             if(mysqli_query($link2, $sql3)){
-
+              echo "temporary modules deleted ";
             } else{
             	   echo "ERROR: Could not able to execute $sql3. " . mysqli_error($link2);
             }
@@ -68,6 +82,6 @@
     }
 
     mysqli_close($link);
-
+  }
 
 ?>
